// src/domain/models/IPerson.ts
export interface IPerson {
  id: string
  nombre: string
  altura: string
  masa: string
  color_de_cabello: string
  color_de_piel: string
  color_de_ojos: string
  año_de_nacimiento: string
  genero: string
  planeta_natal: string
  peliculas: string[]
  especies: string[]
  vehículos: string[]
  naves_estelares: string[]
  creado: string
  editado: string
  url: string
}